+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.22.0] - 2022-06-27
+
 ### Added
 
 - Support for `tradeName` field on user's organization and `phoneNumber` field on user's cost center